fix(test): bot feedback
Resolve BAM-808
