+++ tests/php/bootstrap.php
@@ -38,3 +38,14 @@ require_once realpath(__DIR__ . '/polyfill.php');
 $loader = require realpath(__DIR__ . '/../../vendor/autoload.php');
 
 Doctrine\Common\Annotations\AnnotationRegistry::registerLoader([$loader, 'loadClass']);
+
+if (!function_exists('loadDependencyInjector')) {
+    // Mock DB manager
+    \Tests\Centreon\DependencyInjector::getInstance()[Centreon\ServiceProvider::CENTREON_DB_MANAGER] =
+        new Centreon\Test\Mock\CentreonDBManagerService;
+
+    function loadDependencyInjector()
+    {
+        return \Tests\Centreon\DependencyInjector::getInstance();
+    }
+}
