+++ tests/php/polyfill.php
@@ -41,54 +41,3 @@ if (!class_exists('gnupg')) {
     }
 
 }
-if (!function_exists('loadDependencyInjector')) {
-
-    class DependencyInjector implements ArrayAccess
-    {
-
-        protected static $instance;
-        protected $container;
-
-        public static function getInstance(): self
-        {
-            if (!static::$instance instanceof DependencyInjector) {
-                static::$instance = new DependencyInjector;
-            }
-
-            return static::$instance;
-        }
-
-        public function offsetSet($offset, $value)
-        {
-            if (is_null($offset)) {
-                $this->container[] = $value;
-            } else {
-                $this->container[$offset] = $value;
-            }
-        }
-
-        public function offsetExists($offset)
-        {
-            return isset($this->container[$offset]);
-        }
-
-        public function offsetUnset($offset)
-        {
-            unset($this->container[$offset]);
-        }
-
-        public function offsetGet($offset)
-        {
-            return isset($this->container[$offset]) ? $this->container[$offset] : null;
-        }
-    }
-
-    // Mock DB manager
-    DependencyInjector::getInstance()[Centreon\ServiceProvider::CENTREON_DB_MANAGER] =
-        new Centreon\Test\Mock\CentreonDBManagerService;
-
-    function loadDependencyInjector()
-    {
-        return DependencyInjector::getInstance();
-    }
-}
